Dialogs

+++ app/src/redux/middleware/database-middleware.js
@@ -4,8 +4,8 @@ import ActionTypes from '../actions/action-types'
 
 export default database => {
     return ({dispatch}) => {
-       
-        database.onProjects = projects => 
+
+        database.onProjects = projects =>
             dispatch(ActionsGenerator.onProjects(projects))
 
         return next => action => {
@@ -17,11 +17,13 @@ export default database => {
                     break
 
                 case ActionTypes.SUBMIT_PROJECT:
-                    database.submitProject(action.name, action.description)
+                    let uid = database.submitProject(action.name, action.description)
+                    dispatch(ActionsGenerator.selectProject(uid))
                     break
 
-                case ActionTypes.DELETE_PROJECT:
-                    database.deleteProject(action.uid)
+                case ActionTypes.DELETE_PROJECT_CONFIRM:
+                    if(action.confirmed)
+                        database.deleteProject(action.uid)
                     break
             }
